src/main/java/uk/co/blackpepper/support/spring/webmvc/ResponseStatusExceptionResolver.java
/*
 * Copyright 2014 Black Pepper Software
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package uk.co.blackpepper.support.spring.webmvc;

import java.io.IOException;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

import org.springframework.web.servlet.ModelAndView;
import org.springframework.web.servlet.handler.AbstractHandlerExceptionResolver;

public class ResponseStatusExceptionResolver extends AbstractHandlerExceptionResolver {

	@Override
	protected ModelAndView doResolveException(HttpServletRequest request, HttpServletResponse response, Object handler,
		Exception exception) {
		
		if (!(exception instanceof ResponseStatusException)) {
			return null;
		}
		
		ResponseStatusException responseStatusException = (ResponseStatusException) exception;
		
		try {
			response.sendError(responseStatusException.getStatus().value());
		}
		catch (IOException sendErrorException) {
			logException(sendErrorException, request);
			return null;
		}
		
		return new ModelAndView();
	}
}
